+++ src/render/ShaderLib.js
@@ -17,6 +17,8 @@ import ShaderContour from './shaders/ShaderContour.js';
 import ShaderPaintUV from './shaders/ShaderPaintUV.js';
 import ShaderBlur from './shaders/ShaderBlur.js';
 import ShaderZSphere from './shaders/ShaderZSphere.js';
+import ShaderRefImage from './shaders/ShaderRefImage.js';
+import ShaderViewport2D from './shaders/ShaderViewport2D.js';
 
 var ShaderLib = [];
 
@@ -36,9 +38,10 @@ ShaderLib[Enums.Shader.MERGE] = ShaderMerge;
 ShaderLib[Enums.Shader.FXAA] = ShaderFxaa;
 ShaderLib[Enums.Shader.CONTOUR] = ShaderContour;
 
-// misc
 ShaderLib[Enums.Shader.PAINTUV] = ShaderPaintUV;
 ShaderLib[Enums.Shader.BLUR] = ShaderBlur;
 ShaderLib[Enums.Shader.ZSPHERE] = ShaderZSphere;
+ShaderLib[Enums.Shader.REF_IMAGE] = ShaderRefImage;
+ShaderLib[Enums.Shader.VIEWPORT2D] = ShaderViewport2D;
 
 export default ShaderLib;